# server/v1/docs/api_docs.yml
...
swagger: "2.0"
info:
  description: "This is a detailed ChapiVentas"
  version: "1.0.0"
  title: "ChapiVentas API"
basePath: "/api"
schemes:
- "http"
consumes:
  - application/json
paths:
...
  /promocion:
    post:
      tags:
      - "Promocion"
      summary: "Crear Promocion"
      description: "Crear promocion con fecha inicial y final"
      operationId: "postPromocion"
      parameters:
      - name: "Promocion"
        in: "body"
        description: "Login object"
        required: true
        schema:
          $ref: "#/definitions/Promocion"

      responses:
        200:
          description: ""
          schema:
            type: "string"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
    get:
      tags:
      - "Promocion"
      summary: "Obtener todas las promociones"
      description: "Obtener todas las promociones"
      operationId: "listUsersGET"
      parameters:
      - name: "PromocionDelete"
        in: "body"
        description: "Id Promocion"
        required: true
        schema:
          $ref: "#/definitions/PromocionID"
      responses:
        200:
          description: "Promocion"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Promocion"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
    delete:
      tags:
      - "Promocion"
      summary: "Eliminar una promoción"
      description: "Eliminar una promoción por ID"
      operationId: "deletePromocion"
      responses:
        200:
          description: "Promocion"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Promocion"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  Promocion:
    type: object
    properties:
      tipo:
        type: string
      descuento:
        type: number
      descripcion:
        type: string
      fechaInicioPromo:
        type: string
      fechaFinalPromo:
        type: string
...
  PromocionID:
    type: object
    properties:
      id:
        type: string
...
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
      index:
        type: integer
      errmsg:
        type: "string"
      op:
        type: "array"
        items:
          type: "string"